+++ lib/src/translator.dart
@@ -2,87 +2,56 @@ import 'dart:convert';
 
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
-import 'package:internationalization/internationalization.dart';
+
+part 'localization_delegate.dart';
 
 class Translator {
-  static Translator _instance;
-  static final Map<String, dynamic> _defaultLocaleStrings = new Map();
-
   final String _translationsPath;
   final Locale _locale;
-  final List<String> _files;
+  final Map<String, dynamic> _translations = {};
 
-  Translator._(this._translationsPath, this._locale, this._files);
-
-  Map<String, dynamic> _locationStrings;
-
-  factory Translator.newInstance(String translationsPath, Locale locale, List<String> files) {
-    _instance = Translator._(
-      translationsPath,
-      locale,
-      files
-    );
-
-    return _instance;
+  Translator._(
+    this._locale,
+    this._translationsPath,
+    Map<String, dynamic> externalTranslations,
+  ) {
+    _translations.addAll(externalTranslations);
   }
 
-  static Translator get instance {
-    try {
-      return Localizations.of<Translator>(
-        Internationalization.context,
-        Translator,
-      );
-    } catch (_) {}
-
-    return _instance;
+  void addExternalTranslations(Map<String, dynamic> translations) {
+    _translations.addAll(translations);
   }
 
-  Future<void> load() async {
-    final fullPath = _fullPathJsonStrings(this._locale);
+  static Translator? of(BuildContext context) {
+    return Localizations.of<Translator>(context, Translator);
+  }
 
-    final data = <String>[];
+  Future<void> _load() async {
+    final filePath = _pathJsonStrings(this._locale);
+    final data = await rootBundle.loadString(filePath);
+    final translations = json.decode(data);
+    _translations.addAll(translations);
+  }
 
-    for (String file in _files) {
-      final json = (await rootBundle.loadString("$fullPath$file.json")).trim();
-      final content = json.substring(1, json.length - 1);
-      data.add(content);
+  String _pathJsonStrings(Locale locale) {
+    if (locale.countryCode?.isEmpty != false) {
+      return '$_translationsPath${locale.languageCode.toLowerCase()}.json';
     }
 
-    final jsonContent = data.join(',');
-    _locationStrings = json.decode("{$jsonContent}");
-  }
-
-  String _fullPathJsonStrings(Locale locale) {
-    if (locale?.countryCode?.isEmpty != false) {
-      return '$_translationsPath${locale.languageCode}/';
-    }
-
-    return '$_translationsPath${locale.languageCode}/${locale.countryCode}/';
-  }
-
-  bool _stringExists(String key) =>
-      _locationStrings?.containsKey(key) == true ||
-      _defaultLocaleStrings?.containsKey(key) == true;
-
-  dynamic _valueOf(String key) {
-    if (_locationStrings == null) {
-      return _defaultLocaleStrings[key];
-    }
-
-    return _locationStrings[key] ?? _defaultLocaleStrings[key];
+    return '$_translationsPath${locale.languageCode.toLowerCase()}/${locale.countryCode!.toLowerCase()}.json';
   }
 
   String _interpolateValue(
     String value,
-    List<String> args,
-    Map<String, dynamic> namedArgs,
+    List<String>? args,
+    Map<String, dynamic>? namedArgs,
   ) {
     for (int i = 0; i < (args?.length ?? 0); i++) {
-      value = value.replaceAll("{$i}", args[i]);
+      value = value.replaceAll("{$i}", args![i]);
     }
 
     if (namedArgs?.isNotEmpty == true) {
-      namedArgs.forEach(
+      namedArgs!.forEach(
         (entryKey, entryValue) => value = value.replaceAll(
           "::$entryKey::",
           entryValue.toString(),
@@ -93,37 +62,61 @@ class Translator {
     return value;
   }
 
-  String valueOf(
-    String key, {
-    List<String> args,
-    Map<String, dynamic> namedArgs,
-  }) {
-    if (!_stringExists(key)) {
-      return key;
+  String _valueOf(
+    String? translationContext,
+    String key,
+    List<String>? args,
+    Map<String, dynamic>? namedArgs,
+  ) {
+    final value = _translations[key] ??
+        _translations[translationContext]?[key]?.toString();
+
+    if (value != null) {
+      return _interpolateValue(value, args, namedArgs);
     }
 
-    String value = _valueOf(key).toString();
-
-    value = _interpolateValue(value, args, namedArgs);
-
-    return value;
+    return key;
   }
 
-  String pluralOf(
+  String _pluralOf(
+    String? translationContext,
     String key,
-    int pluralValue, {
-    List<String> args,
-    Map<String, dynamic> namedArgs,
-  }) {
-    if (!_stringExists(key)) {
-      return key;
-    }
+    int pluralValue,
+    List<String>? args,
+    Map<String, dynamic>? namedArgs,
+  ) {
+    final Map<String, dynamic> plurals =
+        _translations[key] ?? _translations[translationContext][key];
 
-    Map<String, dynamic> plurals = _valueOf(key);
     final plural = {0: "zero", 1: "one"}[pluralValue] ?? "other";
     String value = plurals[plural].toString();
     value = _interpolateValue(value, args, namedArgs);
 
     return value;
+  }
+
+  String translate(
+    String key, {
+    String? translationContext,
+    int? pluralValue,
+    List<String>? args,
+    Map<String, dynamic>? namedArgs,
+  }) {
+    if (pluralValue != null) {
+      return _pluralOf(
+        translationContext,
+        key,
+        pluralValue,
+        args,
+        namedArgs,
+      );
+    }
+
+    return _valueOf(
+      translationContext,
+      key,
+      args,
+      namedArgs,
+    );
   }
 }